Extract CLI logic to internal/cli package
Move CLI implementation from cmd/ocw/main.go to internal/cli:
- internal/cli/cli.go: Main CLI logic with Run() and Execute()
- internal/cli/args.go: Argument parsing and reordering
- internal/cli/discovery.go: Workflow file discovery and job lookup
- internal/cli/output.go: User-facing output formatting

Update cmd/ocw/main.go to be a thin entry point (~20 lines) that
delegates to cli.Run(). This separation makes the CLI logic testable
and keeps cmd/ as pure initialization.
